requirements.txt
faiss-cpu
sentence-transformers
pdfplumber
python-docx
pytesseract
Pillow
openai
